+++ packages/react-google-maps-api-gatsby-example/gatsby-node.js
@@ -5,3 +5,11 @@
  */
 
 // You can delete this file if you're not using it
+exports.onCreateBabelConfig = ({ actions }) => {
+  actions.setBabelPreset({
+    name: 'babel-preset-gatsby',
+    options: {
+      reactRuntime: 'automatic',
+    },
+  })
+}