+++ packages/react-google-maps-api-gatsby-example/src/examples/example-ground.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react'
+import { type CSSProperties, memo, useMemo } from 'react'
 import PropTypes from 'prop-types'
 import { GoogleMap, GroundOverlay } from '@react-google-maps/api'
 
@@ -8,23 +8,44 @@ const ExampleGroundPropTypes = {
   }).isRequired,
 }
 
-const center = {
+const center: google.maps.LatLngLiteral = {
   lat: 40.74,
   lng: -74.18,
 }
 
-const BOUNDS = {
+const BoundsLiteral: google.maps.LatLngBoundsLiteral = {
   north: 40.773941,
   south: 40.712216,
   east: -74.12544,
   west: -74.22655,
 }
 
-const onClick = (...args) => {
+const onClick = (...args: any[]) => {
   console.log('onClick args: ', args)
 }
 
-function ExampleGround({ styles }) {
+interface Props {
+  styles: {
+    container: CSSProperties | undefined
+  }
+}
+
+function GroundOverlayC(): JSX.Element {
+  const BOUNDS = useMemo(() => {
+    return new google.maps.LatLngBounds().union(BoundsLiteral)
+  }, [])
+
+  return (
+    <GroundOverlay
+    url='https://www.lib.utexas.edu/maps/historical/newark_nj_1922.jpg'
+    bounds={BOUNDS}
+  />
+  )
+}
+
+const GroundOverlayComponent = memo(GroundOverlayC)
+
+function ExampleGround({ styles }: Props): JSX.Element{
   return (
     <div className='map'>
       <div className='map-container'>
@@ -35,10 +56,7 @@ function ExampleGround({ styles }) {
           center={center}
           onClick={onClick}
         >
-          <GroundOverlay
-            url='https://www.lib.utexas.edu/maps/historical/newark_nj_1922.jpg'
-            bounds={BOUNDS}
-          />
+          <GroundOverlayComponent />
         </GoogleMap>
       </div>
     </div>
@@ -47,4 +65,4 @@ function ExampleGround({ styles }) {
 
 ExampleGround.propTypes = ExampleGroundPropTypes
 
-export default React.memo(ExampleGround)
+export default memo(ExampleGround)