+++ packages/react-google-maps-api-gatsby-example/src/examples/example-overlay-view.stories.tsx
@@ -1,5 +1,4 @@
-import React from 'react'
-import { ComponentStory, ComponentMeta } from '@storybook/react'
+import type { ComponentStory, ComponentMeta } from '@storybook/react'
 import { OverlayView as ExampleComponent} from '@react-google-maps/api'
 
 import ExampleOverlayView from './example-overlay-view'