+++ packages/react-google-maps-api/CHANGELOG.md
@@ -1,5 +1,46 @@
 # Changelog
 
+## 2.10.0 react@18 support, new functional versions of components
+
+- TrafficLayerF
+- BicyclingLayerF
+- TransitLayerF
+- DrawingManagerF
+- MarkerF
+- InfoBoxF
+- InfoWindowF
+- PolylineF
+- PolygonF
+- RectangleF
+- CircleF
+
+This components currently for tests only - I've planned to replace PureComponent class based components in 3.0.0 version.
+
+please test in your applications each F component.
+
+TODO:
+
+- LoadScript
+- GoogleMap
+- MarkerClustererF
+- DirectionsRendererF
+- DirectionsServiceF
+- DistanceMatrixServiceF
+- OverlayViewF
+- HeatmapLayerF
+- KmlLayerF
+- AutocompleteF
+- StandaloneSearchBoxF
+- StreetViewPanoramaF
+- StreetViewServiceF
+
+updated typescript 4.7.0-dev
+installed lerna
+
+## 2.8.1 react@17 support
+
+## 2.7.0 update typescript and dependencies
+
 ## 2.6.0 "@googlemaps/markerclusterer": "1.0.8" GoogleMapsMarkerClusterer component, merged [https://github.com/JustFly1984/react-google-maps-api/pull/2895]( https://github.com/JustFly1984/react-google-maps-api/pull/2895) PR
 
 ## 2.5.0 useJsApiLoader before this patch sets loaded state after loader has loaded, even if the component has already unmounted. This should fixes it